Pr comment fixes
