+++ API.md
@@ -1,4 +1,4 @@
- # Party Service API
+# Party Service API
 
 ## Business Endpoints
 
@@ -1000,7 +1000,7 @@ total would be 27
 ### Delete Respondent by Email
 
 * `DELETE /party-api/v1/respondents/email`
-  * Deletes a respondent by their email address. 
+  * Deletes a respondent by their email address.
   * Returns a 204 on success, a 400 if the id isn't a valid uuid and 404 if the user cannot be found.
 
 ---
@@ -1008,11 +1008,10 @@ total would be 27
 ### Disable all user enrolments by email
 
 * `PUT /party-api/v1/respondents/disable-user-enrolments`
-  * Disables all active enrolments for a respondent based on their email address. 
+  * Disables all active enrolments for a respondent based on their email address.
   * Returns a 204 on success, a 400 if the id isn't a valid uuid and 404 if the user cannot be found.
 
 ---
-
 
 ### Get Respondent Info by Email
