+++ ras_party/controllers/respondent_controller.py
@@ -2,14 +2,13 @@ import logging
 import uuid
 
 import structlog
-from werkzeug.exceptions import BadRequest, NotFound, UnprocessableEntity
-from sqlalchemy.orm.exc import NoResultFound, MultipleResultsFound
+from werkzeug.exceptions import BadRequest, NotFound
 
-from ras_party.controllers.account_controller import change_respondent
+from ras_party.controllers.account_controller import change_respondent, get_single_respondent_by_email
 from ras_party.models.models import Enrolment, BusinessRespondent, PendingEnrolment, Respondent
 from ras_party.controllers.queries import query_respondent_by_party_uuid, \
     query_respondent_by_email, update_respondent_details, query_respondent_by_names_and_emails, \
-    query_respondent_by_party_uuids, query_single_respondent_by_email
+    query_respondent_by_party_uuids
 from ras_party.support.session_decorator import with_db_session, with_query_only_db_session
 from ras_party.support.util import obfuscate_email
 
@@ -94,23 +93,6 @@ def delete_respondent_by_email(email, session):
                 email=obfuscate_email(email),
                 party_uuid=str(respondent.party_uuid),
                 id=respondent.id)
-
-
-def get_single_respondent_by_email(email, session):
-    """gets a single respondent based on an email address"""
-    try:
-        respondent = query_single_respondent_by_email(email, session)
-    except NoResultFound:
-        logger.error("Respondent with email does not exist", email=obfuscate_email(email))
-        raise NotFound("Respondent with email does not exist")
-    except MultipleResultsFound:
-        logger.error("Multiple respondents found for email", email=obfuscate_email(email))
-        raise UnprocessableEntity("Multiple users found, unable to proceed")
-    logger.info("Found respondent",
-                email=obfuscate_email(respondent.email_address),
-                party_uuid=respondent.party_uuid,
-                id=respondent.id)
-    return respondent
 
 
 @with_query_only_db_session
